chore: remove unused legacy infra
