+++ api/src/api/admin.rs
@@ -1,8 +1,8 @@
 // Copyright 2024 the JSR authors. All rights reserved. MIT license.
 use crate::NpmUrl;
 use crate::RegistryUrl;
-use crate::buckets::Buckets;
 use crate::orama::OramaClient;
+use crate::s3::Buckets;
 use hyper::Body;
 use hyper::Request;
 use routerify::Router;
