+++ api/src/gcp.rs
@@ -1,27 +1,14 @@
 // Copyright 2024 the JSR authors. All rights reserved. MIT license.
 use bytes::Bytes;
-use futures::Stream;
-use hyper::HeaderMap;
 use hyper::StatusCode;
-use percent_encoding::NON_ALPHANUMERIC;
-use reqwest::Body;
-use reqwest::Response;
-use reqwest::multipart;
-use reqwest::multipart::Part;
 use serde::Deserialize;
-use serde_json::json;
-use std::borrow::Cow;
 use std::str::FromStr;
 use std::sync::Arc;
 use std::sync::Mutex;
 use std::time::Duration;
 use std::time::Instant;
-use thiserror::Error;
-use tracing::error;
 use tracing::instrument;
 
-pub const CACHE_CONTROL_IMMUTABLE: &str = "public, max-age=31536000, immutable";
-pub const CACHE_CONTROL_DO_NOT_CACHE: &str = "no-cache, no-store, max-age=0";
 const HTTP_CONNECT_TIMEOUT: Duration = Duration::from_secs(10);
 
 #[derive(Deserialize)]
@@ -144,335 +131,6 @@ impl ClientInner {
 }
 
 #[derive(Clone)]
-pub struct Bucket {
-  pub(crate) client: Client,
-  pub(crate) name: String,
-  pub(crate) endpoint: String,
-}
-
-#[derive(Debug, Error, deno_error::JsError)]
-#[class(generic)]
-pub enum GcsError {
-  #[error("request to GCS timed out")]
-  RequestTimeout,
-  #[error("too many requests to GCS")]
-  TooManyRequests,
-  #[error("server error when communicating with GCS: {0} ({0:?})")]
-  Server(StatusCode),
-  #[error("failed to get access token: {0}")]
-  AccessToken(anyhow::Error),
-  #[error(transparent)]
-  Reqwest(#[from] reqwest::Error),
-  #[error("stream failed: {0}")]
-  Stream(anyhow::Error),
-}
-
-impl GcsError {
-  /// 408, 429, and 5xx errors are retryable.
-  /// https://cloud.google.com/storage/docs/retry-strategy
-  pub fn is_retryable(&self) -> bool {
-    matches!(
-      self,
-      Self::RequestTimeout | Self::TooManyRequests | Self::Server(_)
-    )
-  }
-}
-
-#[derive(Debug)]
-pub struct GcsUploadOptions<'a> {
-  pub content_type: Option<Cow<'a, str>>,
-  pub cache_control: Option<Cow<'a, str>>,
-  pub gzip_encoded: bool,
-}
-
-#[allow(dead_code)]
-impl Bucket {
-  pub fn new(client: Client, name: String, endpoint: Option<String>) -> Self {
-    Self {
-      client,
-      name,
-      endpoint: endpoint
-        .unwrap_or_else(|| "https://storage.googleapis.com".to_owned()),
-    }
-  }
-
-  fn error_if_failed(resp: Response) -> Result<Response, GcsError> {
-    let status_code = resp.status();
-    if status_code == StatusCode::REQUEST_TIMEOUT {
-      return Err(GcsError::RequestTimeout);
-    }
-    if status_code == StatusCode::TOO_MANY_REQUESTS {
-      return Err(GcsError::TooManyRequests);
-    }
-    if status_code.is_server_error() {
-      return Err(GcsError::Server(status_code));
-    }
-    let resp = resp.error_for_status()?;
-    Ok(resp)
-  }
-
-  #[cfg(test)]
-  pub async fn create(
-    client: Client,
-    name: String,
-    endpoint: Option<String>,
-  ) -> Result<Self, GcsError> {
-    let bucket = Bucket::new(client, name, endpoint);
-    let url = format!("{}/storage/v1/b", bucket.endpoint);
-    let token = bucket
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let resp = bucket
-      .client
-      .http()
-      .post(&url)
-      .bearer_auth(token)
-      .json(&json!({"name": bucket.name}))
-      .send()
-      .await?;
-    Bucket::error_if_failed(resp)?;
-    Ok(bucket)
-  }
-
-  #[instrument(name = "gcp::Bucket::download_resp", skip(self), err, fields(bucket = %self.name
-  ))]
-  pub async fn download_resp(&self, path: &str) -> Result<Response, GcsError> {
-    let path = percent_encoding::utf8_percent_encode(path, NON_ALPHANUMERIC);
-    let url = format!(
-      "{}/storage/v1/b/{}/o/{}?alt=media",
-      self.endpoint, self.name, path
-    );
-    let token = self
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let resp = self
-      .client
-      .http()
-      .get(url)
-      .bearer_auth(token)
-      .send()
-      .await?;
-    Ok(resp)
-  }
-
-  #[instrument(name = "gcp::Bucket::download", skip(self), err, fields(bucket = %self.name
-  ))]
-  pub async fn download(&self, path: &str) -> Result<Option<Bytes>, GcsError> {
-    let resp = self.download_resp(path).await?;
-    if resp.status() == 404 {
-      return Ok(None);
-    }
-    let resp = Bucket::error_if_failed(resp)?;
-    let bytes = resp.bytes().await?;
-    Ok(Some(bytes))
-  }
-
-  #[instrument(name = "gcp::Bucket::download_stream", skip(self), err, fields(bucket = %self.name
-  ))]
-  pub async fn download_stream(
-    &self,
-    path: &str,
-    offset: Option<usize>,
-  ) -> Result<Option<impl Stream<Item = Result<Bytes, reqwest::Error>>>, GcsError>
-  {
-    self
-      .download_stream_with_encoding(path, offset, "")
-      .await
-      .map(|x| x.map(|x| x.1))
-  }
-
-  #[instrument(
-    name = "gcp::Bucket::download_stream_with_encoding",
-    skip(self),
-    err,
-    fields(bucket = %self.name)
-  )]
-  pub async fn download_stream_with_encoding(
-    &self,
-    path: &str,
-    offset: Option<usize>,
-    accept_encoding: &str,
-  ) -> Result<
-    Option<(
-      HeaderMap,
-      impl Stream<Item = Result<Bytes, reqwest::Error>> + use<>,
-    )>,
-    GcsError,
-  > {
-    let path = percent_encoding::utf8_percent_encode(path, NON_ALPHANUMERIC);
-    let url = format!(
-      "{}/storage/v1/b/{}/o/{}?alt=media",
-      self.endpoint, self.name, path
-    );
-    let token = self
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let mut req = if !accept_encoding.is_empty() {
-      self
-        .client
-        .http_without_compression()
-        .get(url)
-        .header("Accept-Encoding", accept_encoding)
-    } else {
-      self.client.http().get(url)
-    };
-    req = req.bearer_auth(token);
-    if let Some(offset) = offset {
-      // for syntax, refer to https://cloud.google.com/storage/docs/json_api/v1/parameters#offset
-      req = req.header("Range", format!("bytes={offset}-"))
-    }
-    let resp = req.send().await?;
-    if resp.status() == 404 || resp.status() == 416 {
-      return Ok(None);
-    }
-    let mut resp = Bucket::error_if_failed(resp)?;
-    let headers = std::mem::take(resp.headers_mut());
-    Ok(Some((headers, resp.bytes_stream())))
-  }
-
-  async fn upload_inner(
-    &self,
-    path: &str,
-    mut media_part: Part,
-    options: &GcsUploadOptions<'_>,
-  ) -> Result<(), GcsError> {
-    let url = format!(
-      "{}/upload/storage/v1/b/{}/o?uploadType=multipart",
-      self.endpoint, self.name
-    );
-    let token = self
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let request_builder = self.client.http().post(url).bearer_auth(token);
-    let json = json!({
-      "name": path,
-      "cacheControl": options.cache_control,
-      "contentEncoding": if options.gzip_encoded { "gzip" } else { "identity" },
-    });
-    if let Some(content_type) = &options.content_type {
-      media_part = media_part.mime_str(content_type).unwrap();
-    }
-    let meta = Part::text(json.to_string())
-      .mime_str("application/json")
-      .unwrap();
-    let form = multipart::Form::new()
-      .part("meta", meta)
-      .part("media", media_part);
-
-    let resp = request_builder.multipart(form).send().await?;
-    Bucket::error_if_failed(resp)?;
-
-    Ok(())
-  }
-
-  #[instrument(name = "gcp::Bucket::upload", skip(self, data), err, fields(bucket = %self.name, size = %data.len()
-  ))]
-  pub async fn upload(
-    &self,
-    path: &str,
-    data: Bytes,
-    options: &GcsUploadOptions<'_>,
-  ) -> Result<(), GcsError> {
-    self
-      .upload_inner(path, Part::bytes(Cow::Owned(data.to_vec())), options)
-      .await
-  }
-
-  #[instrument(
-    name = "gcp::Bucket::upload_stream",
-    skip(self, stream),
-    err,
-    fields(bucket = %self.name)
-  )]
-  pub async fn upload_stream<
-    S: Stream<Item = Result<Bytes, std::io::Error>> + Send + Sync + 'static,
-  >(
-    &self,
-    path: &str,
-    stream: S,
-    options: &GcsUploadOptions<'_>,
-  ) -> Result<(), GcsError> {
-    self
-      .upload_inner(path, Part::stream(Body::wrap_stream(stream)), options)
-      .await
-  }
-
-  #[instrument(name = "gcp::Bucket::list", skip(self), err, fields(bucket = %self.name
-  ))]
-  pub async fn list(&self, path: &str) -> Result<Option<List>, GcsError> {
-    let path = percent_encoding::utf8_percent_encode(path, NON_ALPHANUMERIC);
-    let url = format!(
-      "{}/storage/v1/b/{}/o?prefix={}",
-      self.endpoint, self.name, path
-    );
-    let token = self
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let resp = self
-      .client
-      .http()
-      .get(url)
-      .bearer_auth(token)
-      .send()
-      .await?;
-
-    if resp.status() == 404 {
-      return Ok(None);
-    }
-    let resp = Bucket::error_if_failed(resp)?;
-    let json = resp.json().await?;
-    Ok(Some(json))
-  }
-
-  #[instrument(name = "gcp::Bucket::delete", skip(self), err, fields(bucket = %self.name
-  ))]
-  pub async fn delete_file(&self, path: &str) -> Result<bool, GcsError> {
-    let path = percent_encoding::utf8_percent_encode(path, NON_ALPHANUMERIC);
-    let url =
-      format!("{}/storage/v1/b/{}/o/{}", self.endpoint, self.name, path);
-    let token = self
-      .client
-      .get_access_token()
-      .await
-      .map_err(GcsError::AccessToken)?;
-    let resp = self
-      .client
-      .http()
-      .delete(url)
-      .bearer_auth(token)
-      .send()
-      .await?;
-
-    if resp.status() == 404 {
-      return Ok(true);
-    }
-    let _ = Bucket::error_if_failed(resp)?;
-    Ok(false)
-  }
-}
-
-#[derive(serde::Deserialize)]
-pub struct List {
-  #[serde(default)]
-  pub items: Vec<ListItem>,
-}
-
-#[derive(serde::Deserialize)]
-pub struct ListItem {
-  pub name: String,
-}
-
-#[derive(Clone)]
 pub struct Queue {
   pub(crate) client: Client,
   pub(crate) id: String,
@@ -521,267 +179,5 @@ impl Queue {
       ));
     }
     Ok(())
-  }
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BigQueryJobReference {
-  pub job_id: String,
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BigQueryQueryResult {
-  pub job_reference: BigQueryJobReference,
-  pub page_token: Option<String>,
-  #[serde(default)]
-  pub rows: Vec<serde_json::Value>,
-  #[serde(default)]
-  pub errors: Option<serde_json::Value>,
-  pub job_complete: bool,
-}
-
-pub struct BigQuery {
-  pub(crate) client: Client,
-  pub(crate) project: String,
-  pub(crate) endpoint: String,
-}
-
-impl BigQuery {
-  pub fn new(
-    client: Client,
-    project: String,
-    endpoint: Option<String>,
-  ) -> Self {
-    Self {
-      client,
-      project,
-      endpoint: endpoint
-        .unwrap_or_else(|| "https://bigquery.googleapis.com/".into()),
-    }
-  }
-
-  #[instrument("gcp::BigQuery::query", skip(self), err, fields(project = %self.project
-  ))]
-  pub async fn query(
-    &self,
-    query: &str,
-    params: &[serde_json::Value],
-  ) -> Result<BigQueryQueryResult, anyhow::Error> {
-    let url = format!(
-      "{}/bigquery/v2/projects/{}/queries",
-      self.endpoint, self.project
-    );
-    let token = self.client.get_access_token().await?;
-    let resp = self
-      .client
-      .http()
-      .post(url)
-      .bearer_auth(token)
-      .json(&json!({ "query": query, "useLegacySql": false, "parameterMode": "NAMED", "queryParameters": params, "formatOptions": { "useInt64Timestamp": true } }))
-      .send()
-      .await?;
-    let status = resp.status();
-    if status != StatusCode::OK {
-      let body = resp.text().await?;
-      return Err(anyhow::anyhow!(
-        "Failed to query BigQuery (status={status}): {body}"
-      ));
-    }
-    let json = resp.json().await?;
-    Ok(json)
-  }
-
-  #[instrument("gcp::BigQuery::get_query_results", skip(self), err, fields(project = %self.project
-  ))]
-  pub async fn get_query_results(
-    &self,
-    job_id: &str,
-    page_token: &str,
-  ) -> Result<BigQueryQueryResult, anyhow::Error> {
-    let url = format!(
-      "{}/bigquery/v2/projects/{}/queries/{}?pageToken={}&formatOptions.useInt64Timestamp=true",
-      self.endpoint, self.project, job_id, page_token
-    );
-    let token = self.client.get_access_token().await?;
-    let resp = self
-      .client
-      .http()
-      .get(url)
-      .bearer_auth(token)
-      .send()
-      .await?;
-    let status = resp.status();
-    if status != StatusCode::OK {
-      let body = resp.text().await?;
-      return Err(anyhow::anyhow!(
-        "Failed to get query results (status={status}): {body}"
-      ));
-    }
-    let json = resp.json().await?;
-    Ok(json)
-  }
-}
-
-/// Fake Google Cloud Storage
-/// https://github.com/fsouza/fake-gcs-server
-///
-/// A single shared server is started on first use and reused across all tests.
-#[cfg(test)]
-pub struct FakeGcsTester {
-  pub port: u16,
-}
-
-#[cfg(test)]
-static SHARED_GCS_PORT: std::sync::OnceLock<u16> = std::sync::OnceLock::new();
-
-#[cfg(test)]
-/// Global list of child processes to kill on exit.
-static TEST_CHILD_PROCS: std::sync::Mutex<Vec<std::process::Child>> =
-  std::sync::Mutex::new(Vec::new());
-
-#[cfg(test)]
-/// Register an atexit handler (once) that kills all tracked child processes.
-static ATEXIT_REGISTERED: std::sync::Once = std::sync::Once::new();
-
-#[cfg(test)]
-pub(crate) fn track_child_proc(proc: std::process::Child) {
-  ATEXIT_REGISTERED.call_once(|| {
-    extern "C" fn kill_children() {
-      if let Ok(mut children) = TEST_CHILD_PROCS.lock() {
-        for child in children.iter_mut() {
-          let _ = child.kill();
-          let _ = child.wait();
-        }
-      }
-    }
-    unsafe extern "C" {
-      fn atexit(cb: extern "C" fn()) -> std::ffi::c_int;
-    }
-    unsafe {
-      atexit(kill_children);
-    }
-  });
-  TEST_CHILD_PROCS.lock().unwrap().push(proc);
-}
-
-#[cfg(test)]
-fn start_shared_fake_gcs() -> u16 {
-  use std::io::BufRead;
-  use std::io::BufReader;
-  use std::process::Stdio;
-
-  #[cfg(all(target_os = "macos", target_arch = "aarch64"))]
-  let p = concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/../tools/bin/darwin-arm64/fake-gcs-server"
-  );
-
-  #[cfg(all(target_os = "macos", target_arch = "x86_64"))]
-  let p = concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/../tools/bin/darwin-amd64/fake-gcs-server"
-  );
-
-  #[cfg(target_os = "linux")]
-  let p = concat!(
-    env!("CARGO_MANIFEST_DIR"),
-    "/../tools/bin/linux-amd64/fake-gcs-server"
-  );
-
-  for attempt in 0..5 {
-    let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-    let port = listener.local_addr().unwrap().port();
-    drop(listener);
-
-    println!("starting shared fake gcs server on port {port}: {p}");
-    let proc = std::process::Command::new(p)
-      .arg(format!("-port={port}"))
-      .arg("-scheme=http")
-      .arg("-backend=memory")
-      .stderr(Stdio::piped())
-      .spawn();
-
-    let mut proc = match proc {
-      Ok(p) => p,
-      Err(e) => {
-        eprintln!("attempt {attempt}: failed to spawn fake gcs server: {e}");
-        continue;
-      }
-    };
-
-    let stderr = proc.stderr.take().unwrap();
-    let mut stderr = BufReader::new(stderr);
-    let mut first_line = String::new();
-    if stderr.read_line(&mut first_line).is_err() {
-      let _ = proc.kill();
-      eprintln!("attempt {attempt}: failed to read from fake gcs server");
-      continue;
-    }
-    if !first_line.contains("server started at http://") {
-      let _ = proc.kill();
-      eprintln!("attempt {attempt}: unexpected fake gcs output: {first_line}");
-      continue;
-    }
-
-    // Drain stderr to prevent blocking.
-    std::thread::spawn(move || {
-      std::io::copy(&mut stderr.into_inner(), &mut std::io::sink()).ok();
-    });
-
-    track_child_proc(proc);
-    return port;
-  }
-  panic!("failed to start shared fake gcs server after 5 attempts");
-}
-
-#[cfg(test)]
-impl FakeGcsTester {
-  pub fn new() -> Self {
-    let port = *SHARED_GCS_PORT.get_or_init(start_shared_fake_gcs);
-    Self { port }
-  }
-
-  pub fn endpoint(&self) -> String {
-    format!("http://localhost:{}", self.port)
-  }
-
-  pub async fn create_bucket(&self, bucket: &str) -> Bucket {
-    let client = Client::new(MetadataStrategy::Testing);
-    Bucket::create(client, bucket.to_owned(), Some(self.endpoint()))
-      .await
-      .unwrap()
-  }
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-
-  #[tokio::test]
-  async fn gcs_upload_download() {
-    let tester = FakeGcsTester::new();
-    let bucket = tester.create_bucket("testbucket").await;
-
-    bucket
-      .upload(
-        "upload_download.txt",
-        "hello world".as_bytes().to_vec().into(),
-        &GcsUploadOptions {
-          content_type: None,
-          cache_control: None,
-          gzip_encoded: false,
-        },
-      )
-      .await
-      .unwrap();
-
-    let response = bucket.download("upload_download.txt").await.unwrap();
-    assert!(response.is_some());
-    assert_eq!(response.unwrap().len(), 11);
-
-    let response = bucket.download("does_not_exist.txt").await.unwrap();
-    assert!(response.is_none());
   }
 }
