+++ api/src/publish.rs
@@ -5,7 +5,6 @@ use std::collections::HashSet;
 use crate::NpmUrl;
 use crate::RegistryUrl;
 use crate::api::ApiError;
-use crate::buckets::Buckets;
 use crate::db::Database;
 use crate::db::DependencyKind;
 use crate::db::ExportsMap;
@@ -17,9 +16,6 @@ use crate::db::PackageVersionMeta;
 use crate::db::PublishingTask;
 use crate::db::PublishingTaskError;
 use crate::db::PublishingTaskStatus;
-use crate::gcp::CACHE_CONTROL_DO_NOT_CACHE;
-use crate::gcp::CACHE_CONTROL_IMMUTABLE;
-use crate::gcp::GcsUploadOptions;
 use crate::ids::PackagePath;
 use crate::metadata::ManifestEntry;
 use crate::metadata::PackageMetadata;
@@ -27,6 +23,10 @@ use crate::metadata::VersionMetadata;
 use crate::npm::NPM_TARBALL_REVISION;
 use crate::npm::generate_npm_version_manifest;
 use crate::orama::OramaClient;
+use crate::s3::Buckets;
+use crate::s3::CACHE_CONTROL_DO_NOT_CACHE;
+use crate::s3::CACHE_CONTROL_IMMUTABLE;
+use crate::s3::S3UploadOptions;
 use crate::s3::UploadTaskBody;
 use crate::tarball::NpmTarballInfo;
 use crate::tarball::ProcessTarballOutput;
@@ -297,7 +297,7 @@ async fn upload_version_manifest(
     .upload(
       version_metadata_gcs_path.into(),
       UploadTaskBody::Bytes(content.into()),
-      GcsUploadOptions {
+      S3UploadOptions {
         content_type: Some("application/json".into()),
         cache_control: Some(CACHE_CONTROL_IMMUTABLE.into()),
         gzip_encoded: false,
@@ -405,7 +405,7 @@ async fn upload_package_manifest(
     .upload(
       package_metadata_gcs_path.into(),
       UploadTaskBody::Bytes(content.into()),
-      GcsUploadOptions {
+      S3UploadOptions {
         content_type: Some("application/json".into()),
         cache_control: Some(CACHE_CONTROL_DO_NOT_CACHE.into()),
         gzip_encoded: false,
@@ -440,7 +440,7 @@ async fn upload_npm_version_manifest(
     .upload(
       npm_version_manifest_path_gcs_path.into(),
       crate::s3::UploadTaskBody::Bytes(content.into()),
-      GcsUploadOptions {
+      S3UploadOptions {
         content_type: Some("application/json".into()),
         cache_control: Some(CACHE_CONTROL_DO_NOT_CACHE.into()),
         gzip_encoded: false,
@@ -530,7 +530,7 @@ pub mod tests {
       .upload(
         tarball_path.into(),
         crate::s3::UploadTaskBody::Bytes(tarball_data),
-        GcsUploadOptions {
+        S3UploadOptions {
           content_type: Some("application/x-tar".into()),
           cache_control: None,
           gzip_encoded: true,
